fix: autofix payload to pass context (#321)
