+++ infrastructure/code/types.go
@@ -26,6 +26,34 @@ type SnykAnalysisFailedError struct {
 
 func (e SnykAnalysisFailedError) Error() string { return e.Msg }
 
+type AnalysisRequestKey struct {
+	Type         string   `json:"type"`
+	Hash         string   `json:"hash"`
+	LimitToFiles []string `json:"limitToFiles,omitempty"`
+	Shard        string   `json:"shard"`
+}
+
+type codeRequestContextOrg struct {
+	Name        string          `json:"name"`
+	DisplayName string          `json:"displayName"`
+	PublicId    string          `json:"publicId"`
+	Flags       map[string]bool `json:"flags"`
+}
+
+type codeRequestContext struct {
+	Initiator string                `json:"initiator"`
+	Flow      string                `json:"flow,omitempty"`
+	Org       codeRequestContextOrg `json:"org,omitempty"`
+}
+
+type AnalysisRequest struct {
+	Key             AnalysisRequestKey `json:"key"`
+	Severity        int                `json:"severity,omitempty"`
+	Prioritized     bool               `json:"prioritized,omitempty"`
+	Legacy          bool               `json:"legacy"`
+	AnalysisContext codeRequestContext `json:"analysisContext"`
+}
+
 // AutofixResponse is the json-based structure to which we can translate the results of the HTTP
 // request to Autofix upstream.
 type AutofixResponse struct {
@@ -50,8 +78,8 @@ type AutofixRequestKey struct {
 }
 
 type AutofixRequest struct {
-	Key            AutofixRequestKey  `json:"key"`
-	AutofixContext codeRequestContext `json:"autofixContext"`
+	Key             AutofixRequestKey  `json:"key"`
+	AnalysisContext codeRequestContext `json:"analysisContext"`
 }
 
 // Should implement `error` interface
@@ -68,6 +96,7 @@ type AutofixSuggestion struct {
 }
 
 type AutofixFeedback struct {
-	FixId    string `json:"fixId"`
-	Feedback string `json:"feedback"`
+	FixId           string             `json:"fixId"`
+	Feedback        string             `json:"feedback"`
+	AnalysisContext codeRequestContext `json:"analysisContext"`
 }
